call mainThreadApi directly, improve types
