+++ client/shared/src/api/contract.ts
@@ -7,6 +7,7 @@ import { ProxySubscribable } from './extension/api/common'
 import { TextDocumentPositionParameters } from './protocol'
 import { MaybeLoadingResult } from '@sourcegraph/codeintellify'
 import { HoverMerged } from './client/types/hover'
+import { GraphQLResult } from '../graphql/graphql'
 
 /**
  * This is exposed from the extension host thread to the main thread
@@ -44,6 +45,14 @@ export interface MainThreadAPI {
      * Applies a settings update from extensions.
      */
     applySettingsEdit: (edit: SettingsEdit) => Promise<void>
+
+    /**
+     * GraphQL request API
+     */
+    requestGraphQL: <TResult extends object = any, TVariables extends object = any>(
+        request: string,
+        variables: TVariables
+    ) => Promise<GraphQLResult<TResult>>
 
     // Commands
     executeCommand: (command: string, args: any[]) => Promise<any>
